tests: shell: add more json dumps
Those are expected to be stable, so add them.
Some are not 100% correct, as "typeof" is misprinted as "type" (json output
and input parser lack support for this), but for these files the "type"
is valid too.

This will allow better validation once proper "typeof" support is
added to json.c and json-parser.c.

Signed-off-by: Florian Westphal <fw@strlen.de>